ai-pipeline/requirements.txt
absl-py==2.3.1
absolufy-imports==0.3.1
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
Authlib==1.6.5
babel==2.17.0
beautifulsoup4==4.14.2
black==25.9.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
build==1.3.0
CacheControl==0.14.3
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cleantext==1.1.4
cleo==2.1.0
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
crashtest==0.4.1
cryptography==46.0.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.4
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dulwich==0.24.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
executing==2.2.1
fastapi==0.118.1
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gdown==5.2.0
google-adk==1.15.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.119.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.3
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.33.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.40.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.1
grpcio-status==1.75.1
gym==0.23.0
gym-notices==0.1.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.2
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.3.0
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
onnxruntime==1.23.1
openai==2.2.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.9.0a0
opentelemetry-exporter-gcp-monitoring==1.9.0a0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pbs-installer==2025.10.7
-e git+https://github.com/google/adk-samples.git@66d7afb29b542bc87d1b9789127bd3b925029b3d#egg=personalized_shopping&subdirectory=python/agents/personalized-shopping
pexpect==4.9.0
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyee==13.0.0
Pygments==2.19.2
pyjnius==1.7.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyserini==0.43.0
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
thefuzz==0.22.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trove-classifiers==2025.9.11.17
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.34.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
xattr==1.2.0
zipp==3.23.0
zstandard==0.25.0
